Moved the System Requirements check from SystemInfoBundle
