+++ bundles/AdminBundle/templates/admin/install/check.html.twig
