+++ bundles/SystemInfoBundle/public/css/icons.css
@@ -14,10 +14,6 @@
     background: url(/bundles/pimcoreadmin/img/flat-color-icons/php.svg) center center no-repeat !important;
 }
 
-.pimcore_icon_systemrequirements {
-    background: url(/bundles/pimcoreadmin/img/flat-color-icons/factory.svg) center center no-repeat !important;
-}
-
 .pimcore_icon_reports {
     background: url(/bundles/pimcoreadmin/img/flat-color-icons/pie_chart.svg) center center no-repeat !important;
 }
