+++ bundles/SystemInfoBundle/translations/admin.en.yaml
@@ -1,3 +1,2 @@
 bundle_systemInfo_php_opcache_status: PHP OPcache status
-bundle_systemInfo_system_requirements_check: System-Requirements Check
 bundle_systemInfo_php_info: PHP Info
